Marcus update
Major revisions: -PlayerProfile: I made this class to keep track of a
player instance's current status
-GameProfile: I made this class to keep track of all things related to a
specific game. It is not being used for much right now, but it stores
the player names, the graph, the seed, the winner, the number of turns a
game took, and a game ID (This is so, the same exact graph and players,
but with different gameID). I figure we could use this info to print out
the results of a match, but it's up to you.
-GameMaster: This class has undergone dramatic changes. I use
PlayerProfiles and GameProfiles to manage the matches and I have a
method that executes a game for one round (which takes a first player
and second player and has them duke it out till the turns are up or they
both have full hands) on a given graph. I also added a runMatches()
method which executes two oneRound() methods but swaps the players, each
round creating a new GameProfile with a unique ID and stores the winners
and their hands. The runMatches() method can run multiple matches of the
same graph if need be (I currently have it set to 1). I also added a
verbosity variable, which is hardcoded to true for now, but we can turn
it off for speed during the tournament.